Improving serialization storage space
Until now we only optimized the long tail, where the serialization
ignored all buckets above the last seen value. This is important,
because in a general system many measurements do not see the high latencies
that are near the maximum.

However the same is true for low latencies in a lot of systems, where
there is absolutely no way that many measured transactions end up
with latencies below 100ms or 200ms, leaving the especially "dense"
starting phase with just tons of zeros.

This small change skips all starting count buckets until there is
the first non-zero one. It remembers the amount of skipped buckets
and during `hdr_import` fills them with zeros before starting to
reconstruct the data.