Improve export/import by moving from hashmap to negative skip counters.
The hdr_export method now returns a key "v" for the values, by using
positive numbers when there are observed values in the histogram and
negative numbers for the skip counter of "empty" buckets to keep at 0.

hdr_import is again kept backwards compatible to allow dealing with
format version 2 or 3.